FAT2-295 - Update provider service remote to call outer api

+++ src/SFA.DAS.EmployerFinance/Infrastructure/OuterApiRequests/GetProviderRequest.cs
@@ -4,9 +4,9 @@ namespace SFA.DAS.EmployerFinance.Infrastructure.OuterApiRequests
 {
     public class GetProviderRequest : IGetApiRequest
     {
-        private readonly int _id;
+        private readonly long _id;
 
-        public GetProviderRequest (int id)
+        public GetProviderRequest (long id)
         {
             _id = id;
         }
